CHANGESET: FS-1834 check for undefined repository in EmojiResource
Summary: FS-1834 check for undefined repository in EmojiResource

Release notes: <none>

Releases: @atlaskit/emoji@patch

Dependents: @atlaskit/conversation@none, @atlaskit/editor-bitbucket-transformer@none, @atlaskit/editor-common@none, @atlaskit/editor-core@none, @atlaskit/editor-jira-transformer@none, @atlaskit/reactions@none, @atlaskit/comment@none, @atlaskit/editor-confluence-transformer@none, @atlaskit/editor-json-transformer@none, @atlaskit/editor-markdown-transformer@none, @atlaskit/editor-test-helpers@none, @atlaskit/renderer@none, @atlaskit/task-decision@none

---
{"summary":"FS-1834 check for undefined repository in EmojiResource","releases":[{"name":"@atlaskit/emoji","type":"patch"}],"dependents":[{"name":"@atlaskit/conversation","dependencies":["@atlaskit/emoji","@atlaskit/reactions","@atlaskit/editor-core","@atlaskit/renderer","@atlaskit/comment","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/editor-bitbucket-transformer","dependencies":["@atlaskit/emoji","@atlaskit/editor-core","@atlaskit/editor-test-helpers","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/editor-common","dependencies":["@atlaskit/emoji","@atlaskit/editor-test-helpers"],"type":"none"},{"name":"@atlaskit/editor-core","dependencies":["@atlaskit/emoji","@atlaskit/renderer","@atlaskit/task-decision","@atlaskit/editor-test-helpers","@atlaskit/editor-markdown-transformer","@atlaskit/editor-json-transformer","@atlaskit/editor-common","@atlaskit/editor-bitbucket-transformer"],"type":"none"},{"name":"@atlaskit/editor-jira-transformer","dependencies":["@atlaskit/emoji","@atlaskit/editor-core","@atlaskit/editor-test-helpers","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/reactions","dependencies":["@atlaskit/emoji"],"type":"none"},{"name":"@atlaskit/comment","dependencies":["@atlaskit/editor-core"],"type":"none"},{"name":"@atlaskit/editor-confluence-transformer","dependencies":["@atlaskit/editor-core","@atlaskit/editor-test-helpers","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/editor-json-transformer","dependencies":["@atlaskit/editor-core","@atlaskit/editor-test-helpers","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/editor-markdown-transformer","dependencies":["@atlaskit/editor-core","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/editor-test-helpers","dependencies":["@atlaskit/editor-core","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/renderer","dependencies":["@atlaskit/editor-core","@atlaskit/task-decision","@atlaskit/editor-test-helpers","@atlaskit/editor-json-transformer","@atlaskit/editor-common"],"type":"none"},{"name":"@atlaskit/task-decision","dependencies":["@atlaskit/renderer","@atlaskit/editor-common"],"type":"none"}]}
---

+++ packages/elements/emoji/src/api/EmojiResource.ts
@@ -241,7 +241,7 @@ export class EmojiResource extends AbstractResource<
   SearchOptions
 > implements EmojiProvider {
   protected recordConfig?: ServiceConfig;
-  protected emojiRepository: EmojiRepository;
+  protected emojiRepository?: EmojiRepository;
   protected lastQuery: LastQuery;
   protected activeLoaders: number = 0;
   protected retries: Map<Retry<any>, ResolveReject<any>> = new Map();
@@ -264,10 +264,10 @@ export class EmojiResource extends AbstractResource<
       const emojis = loader.loadEmoji();
       emojis
         .then(emojiResponse => {
-          this.activeLoaders--;
           emojiResponses[index] = emojiResponse;
           this.initEmojiRepository(emojiResponses);
           this.initSiteEmojiResource(emojiResponse, provider).then(() => {
+            this.activeLoaders--;
             this.performRetries();
             this.refreshLastFilter();
           });
@@ -432,7 +432,7 @@ export class EmojiResource extends AbstractResource<
   findByShortName(
     shortName: string,
   ): OptionalEmojiDescription | Promise<OptionalEmojiDescription> {
-    if (this.isLoaded()) {
+    if (this.emojiRepository) {
       // Wait for all emoji to load before looking by shortName (to ensure correct priority)
       return this.emojiRepository.findByShortName(shortName);
     }
@@ -497,14 +497,14 @@ export class EmojiResource extends AbstractResource<
   }
 
   getAsciiMap(): Promise<Map<string, EmojiDescription>> {
-    if (this.isLoaded()) {
+    if (this.emojiRepository) {
       return Promise.resolve(this.emojiRepository.getAsciiMap());
     }
     return this.retryIfLoading(() => this.getAsciiMap(), new Map());
   }
 
   getFrequentlyUsed(options?: SearchOptions): Promise<EmojiDescription[]> {
-    if (this.isLoaded()) {
+    if (this.emojiRepository) {
       return Promise.resolve(this.emojiRepository.getFrequentlyUsed(options));
     }
 
@@ -544,7 +544,7 @@ export class EmojiResource extends AbstractResource<
   deleteSiteEmoji(emoji: EmojiDescription): Promise<boolean> {
     if (this.siteEmojiResource && emoji.id) {
       return this.siteEmojiResource.deleteEmoji(emoji).then(success => {
-        if (success) {
+        if (this.emojiRepository && success) {
           this.emojiRepository.delete(emoji);
         }
         return success;
@@ -573,7 +573,7 @@ export class EmojiResource extends AbstractResource<
   }
 
   calculateDynamicCategories(): Promise<string[]> {
-    if (this.isLoaded()) {
+    if (this.emojiRepository) {
       return Promise.resolve(this.emojiRepository.getDynamicCategoryList());
     }
 
@@ -585,7 +585,9 @@ export class EmojiResource extends AbstractResource<
   }
 
   protected addUnknownEmoji(emoji: EmojiDescription) {
-    this.emojiRepository.addUnknownEmoji(emoji);
+    if (this.emojiRepository) {
+      this.emojiRepository.addUnknownEmoji(emoji);
+    }
   }
 }
